Finished day 21.
Life lesson: Don't casually change the name of a variable to fix a
naming collision. Both of the bugs in my original implementation were
due to changing the variable 'armor' to 'armor_set' and failing to
update all locations.

I even had a thought that I should audit all uses of 'armor', but
foolishly didn't and spent a long time debugging instead.